app/requirements.txt
requests2==2.16.0
selenium==3.141.0
Flask==1.1.2
flask_expects_json==1.4.0